Create README.md
The Reset Database is an open source research project designed to make the recruiting process more efficient.

The goal is to create a pipeline between qualified candidates from the hidden talent market and the practical small businesses whose recruiting resources do not have the reach of larger corporations. The first step is to identify that talent by finding out who's not getting interviews and why not.